+++ CHANGELOG.md
@@ -28,6 +28,7 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the `Unr
 
 -   [#129](https://github.com/FC4E-WP5/fc4eosc-PIDMR-api/pull/129) - PIDMR-206 Add validator Property to Provider Table.
 -   [#133](https://github.com/FC4E-WP5/fc4eosc-PIDMR-api/pull/133) - PIDMR-225 PIDMR provider schema update.
+  - [#138](https://github.com/FC4E-WP5/fc4eosc-PIDMR-api/pull/138) PIDMR-238 Create Endpoint for Getting a Provider by ID for Users
 
 
 ## 2.3.0 - 2024-12-19